impl/pgvector/requirements.txt
psycopg2-binary>=2.9.0
sentence-transformers>=2.2.0
anthropic>=0.18.0
python-dotenv>=1.0.0
